feat: Implement new Modal for deposit
