+++ vite.config.ts
@@ -7,7 +7,8 @@ export default defineConfig({
   plugins: [react()],
   resolve: {
     alias: {
-      "@": path.resolve(import.meta.dirname, "./src"),
+      "@": path.resolve(import.meta.dirname, "src"),
+      "@components": path.resolve(__dirname, "src/components")
     },
   },
   css: {
